warehouse: Rename service directory
